requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
gunicorn==20.1.0
openai==0.27.0
python-dotenv==0.19.0
spotipy==2.23.0
Werkzeug==2.0.1
requests==2.26.0
urllib3==1.26.7